ecommerce/requirements.txt
asgiref==3.11.1
boto3==1.42.94
botocore==1.42.94
certifi==2026.2.25
charset-normalizer==3.4.7
Django==4.2.29
django-cors-headers==4.9.0
django-countries==8.2.0
django-elasticsearch-dsl==8.2
django-phonenumber-field==8.3.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
elastic-transport==8.17.1
elasticsearch==8.19.3
elasticsearch-dsl==8.18.0
idna==3.11
jmespath==1.1.0
phonenumbers==9.0.26
pillow==11.3.0
psycopg2==2.9.12
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.1
six==1.17.0
sqlparse==0.5.5
stripe==15.1.0
typing_extensions==4.15.0
urllib3==1.26.20
